+++ library/Dlayer/DesignerTool/ContentManager/Html/Tool.php
@@ -172,8 +172,8 @@ class Dlayer_DesignerTool_ContentManager_Html_Tool extends Dlayer_Tool_Content
      */
     protected function validateInstances($site_id, $content_id)
     {
-        $model = new Dlayer_DesignerTool_ContentManager_Html_Model();
-        $instances = $model->instancesOfData($site_id, $content_id);
+        $model = new Dlayer_DesignerTool_ContentManager_Shared_Model_Content();
+        $instances = $model->instancesOfData($site_id, $content_id, 'Html');
 
         if($instances > 1)
         {