+++ library/Dlayer/DesignerTool/ContentManager/Jumbotron/Tool.php
@@ -73,8 +73,8 @@ class Dlayer_DesignerTool_ContentManager_Jumbotron_Tool extends Dlayer_Tool_Cont
 	 */
 	protected function validateInstances($site_id, $content_id)
 	{
-		$model_jumbotron = new Dlayer_DesignerTool_ContentManager_Jumbotron_Model();
-		$instances = $model_jumbotron->instancesOfData($site_id, $content_id);
+        $model = new Dlayer_DesignerTool_ContentManager_Shared_Model_Content();
+        $instances = $model->instancesOfData($site_id, $content_id, 'Jumbotron');
 
 		if($instances > 1)
 		{