+++ library/Dlayer/DesignerTool/ContentManager/Text/Ribbon.php
@@ -70,17 +70,7 @@ class Dlayer_DesignerTool_ContentManager_Text_Ribbon extends Dlayer_Ribbon_Conte
 	 */
 	protected function instancesOfData()
 	{
-	    if ($this->instances_of_fetched === false) {
-
-	        $this->instances_of = 0;
-
-            if ($this->tool['content_id'] !== null) {
-                $model = new Dlayer_DesignerTool_ContentManager_Text_Model();
-                $this->instances_of = $model->instancesOfData($this->tool['site_id'], $this->tool['content_id']);
-            }
-
-            $this->instances_of_fetched = true;
-        }
+        $this->instancesOfDataForBaseTool('Text');
 	}
 
     /**